+++ src/components/shared/Game/GameSummary/GameSummaryModal/GameSummaryHistoryRevealedPlayerLine.vue
@@ -1,15 +1,15 @@
 <template>
     <tr class="row">
-        <td class="col-1 d-flex flex-column align-items-center justify-content-center">
+        <td class="d-none d-lg-flex col-lg-1 flex-column align-items-center justify-content-center">
             <img width="45" alt="Seen Icon" :src="SVGs.seen"/>
         </td>
-        <td class="col-4 d-flex flex-column justify-content-center align-items-center">
+        <td class="col-6 col-lg-4 d-flex flex-column justify-content-center align-items-center">
             <RoleImage width="30" :role="revealedPlayer.role.current"/>
             <div class="text-center" v-html="revealedPlayer.name"/>
             <RoleText class="text-center small cursor-text" :role="revealedPlayer.role.current"/>
         </td>
-        <td class="col-7 d-flex justify-content-center align-items-center">
-            <span v-html="revealedText"/>
+        <td class="col-6 col-lg-7 d-flex justify-content-center align-items-center">
+            <span class="text-center" v-html="revealedText"/>
         </td>
     </tr>
 </template>